@@ -1,4 +1,4 @@
-// +build linux,amd64
+//go:build linux && amd64
 
 package gqlmgmt
 
