@@ -15,24 +15,6 @@ type CArgs struct {
 	strMems []unsafe.Pointer // C strings
 }
 
-// NewCArgs constructs CArgs.
-func NewCArgs(args []string) *CArgs {
-	a := new(CArgs)
-	a.Argc = len(args)
-
-	var b *C.char
-	ptrSize := unsafe.Sizeof(b)
-	a.Argv = C.malloc(C.size_t(ptrSize) * C.size_t(len(args)))
-
-	for i, arg := range args {
-		argvEle := (**C.char)(unsafe.Pointer(uintptr(a.Argv) + uintptr(i)*ptrSize))
-		*argvEle = C.CString(arg)
-		a.strMems = append(a.strMems, unsafe.Pointer(*argvEle))
-	}
-
-	return a
-}
-
 // Close releases C memory in CArgs.
 func (a *CArgs) Close() error {
 	for _, strMem := range a.strMems {
@@ -40,4 +22,22 @@ func (a *CArgs) Close() error {
 	}
 	C.free(unsafe.Pointer(a.Argv))
 	return nil
+}
+
+// NewCArgs constructs CArgs.
+func NewCArgs(args []string) *CArgs {
+	a := new(CArgs)
+	a.Argc = len(args)
+
+	var b *C.char
+	ptrSize := int(unsafe.Sizeof(b))
+	a.Argv = C.malloc(C.size_t(ptrSize) * C.size_t(len(args)))
+
+	for i, arg := range args {
+		argvEle := (**C.char)(unsafe.Add(a.Argv, i*ptrSize))
+		*argvEle = C.CString(arg)
+		a.strMems = append(a.strMems, unsafe.Pointer(*argvEle))
+	}
+
+	return a
 }
