@@ -12,6 +12,7 @@ import "C"
 import (
 	"errors"
 	"io"
+	"runtime/cgo"
 	"unsafe"
 
 	"github.com/usnistgov/ndn-dpdk/core/cptr"
@@ -115,11 +116,11 @@ func (bd *Bdev) ReadBlocks(blockOffset, blockCount int64, buf []byte) error {
 	defer eal.Free(bufC)
 
 	done := make(chan error)
-	ctx := cptr.CtxPut(done)
-	defer cptr.CtxClear(ctx)
+	ctx := cgo.NewHandle(done)
+	defer ctx.Delete()
 	eal.PostMain(cptr.Func0.Void(func() {
 		res := C.spdk_bdev_read_blocks(bd.c, bd.ch, bufC, C.uint64_t(blockOffset), C.uint64_t(blockCount),
-			C.spdk_bdev_io_completion_cb(C.go_bdevIoComplete), ctx)
+			C.spdk_bdev_io_completion_cb(C.go_bdevIoComplete), unsafe.Pointer(ctx))
 		if res != 0 {
 			done <- eal.MakeErrno(res)
 		}
@@ -147,11 +148,11 @@ func (bd *Bdev) WriteBlocks(blockOffset, blockCount int64, buf []byte) error {
 	C.rte_memcpy(bufC, unsafe.Pointer(&buf[0]), C.size_t(sizeofBuf))
 
 	done := make(chan error)
-	ctx := cptr.CtxPut(done)
-	defer cptr.CtxClear(ctx)
+	ctx := cgo.NewHandle(done)
+	defer ctx.Delete()
 	eal.PostMain(cptr.Func0.Void(func() {
 		res := C.spdk_bdev_write_blocks(bd.c, bd.ch, bufC, C.uint64_t(blockOffset), C.uint64_t(blockCount),
-			C.spdk_bdev_io_completion_cb(C.go_bdevIoComplete), ctx)
+			C.spdk_bdev_io_completion_cb(C.go_bdevIoComplete), unsafe.Pointer(ctx))
 		if res != 0 {
 			done <- eal.MakeErrno(res)
 		}
@@ -162,11 +163,11 @@ func (bd *Bdev) WriteBlocks(blockOffset, blockCount int64, buf []byte) error {
 // UnmapBlocks notifies the device that the data in the blocks are no longer needed.
 func (bd *Bdev) UnmapBlocks(blockOffset, blockCount int64) error {
 	done := make(chan error)
-	ctx := cptr.CtxPut(done)
-	defer cptr.CtxClear(ctx)
+	ctx := cgo.NewHandle(done)
+	defer ctx.Delete()
 	eal.PostMain(cptr.Func0.Void(func() {
 		res := C.spdk_bdev_unmap_blocks(bd.c, bd.ch, C.uint64_t(blockOffset), C.uint64_t(blockCount),
-			C.spdk_bdev_io_completion_cb(C.go_bdevIoComplete), ctx)
+			C.spdk_bdev_io_completion_cb(C.go_bdevIoComplete), unsafe.Pointer(ctx))
 		if res != 0 {
 			done <- eal.MakeErrno(res)
 		}
@@ -177,12 +178,12 @@ func (bd *Bdev) UnmapBlocks(blockOffset, blockCount int64) error {
 // ReadPacket reads blocks via scatter gather list.
 func (bd *Bdev) ReadPacket(blockOffset, blockCount int64, pkt pktmbuf.Packet) error {
 	done := make(chan error)
-	ctx := cptr.CtxPut(done)
-	defer cptr.CtxClear(ctx)
+	ctx := cgo.NewHandle(done)
+	defer ctx.Delete()
 	eal.PostMain(cptr.Func0.Void(func() {
 		res := C.SpdkBdev_ReadPacket(bd.c, bd.ch, (*C.struct_rte_mbuf)(pkt.Ptr()),
 			C.uint64_t(blockOffset), C.uint64_t(blockCount), C.uint32_t(bd.blockSize),
-			C.spdk_bdev_io_completion_cb(C.go_bdevIoComplete), ctx)
+			C.spdk_bdev_io_completion_cb(C.go_bdevIoComplete), unsafe.Pointer(ctx))
 		if res != 0 {
 			done <- eal.MakeErrno(res)
 		}
@@ -193,12 +194,12 @@ func (bd *Bdev) ReadPacket(blockOffset, blockCount int64, pkt pktmbuf.Packet) er
 // WritePacket writes blocks via scatter gather list.
 func (bd *Bdev) WritePacket(blockOffset, blockCount int64, pkt pktmbuf.Packet) error {
 	done := make(chan error)
-	ctx := cptr.CtxPut(done)
-	defer cptr.CtxClear(ctx)
+	ctx := cgo.NewHandle(done)
+	defer ctx.Delete()
 	eal.PostMain(cptr.Func0.Void(func() {
 		res := C.SpdkBdev_WritePacket(bd.c, bd.ch, (*C.struct_rte_mbuf)(pkt.Ptr()),
 			C.uint64_t(blockOffset), C.uint64_t(blockCount), C.uint32_t(bd.blockSize),
-			C.spdk_bdev_io_completion_cb(C.go_bdevIoComplete), ctx)
+			C.spdk_bdev_io_completion_cb(C.go_bdevIoComplete), unsafe.Pointer(ctx))
 		if res != 0 {
 			done <- eal.MakeErrno(res)
 		}
@@ -208,7 +209,7 @@ func (bd *Bdev) WritePacket(blockOffset, blockCount int64, pkt pktmbuf.Packet) e
 
 //export go_bdevIoComplete
 func go_bdevIoComplete(io *C.struct_spdk_bdev_io, success C.bool, ctx unsafe.Pointer) {
-	done := cptr.CtxGet(ctx).(chan error)
+	done := cgo.Handle(ctx).Value().(chan error)
 	if bool(success) {
 		done <- nil
 	} else {
