@@ -9,6 +9,13 @@ import (
 	"github.com/graphql-go/graphql"
 )
 
+// Variables replaced via -ldflags -X.
+var (
+	commit string
+	date   string
+	dirty  string
+)
+
 // Version records NDN-DPDK version information.
 type Version struct {
 	Version string    `json:"version"`
@@ -23,7 +30,7 @@ func (v Version) String() string {
 
 // Get returns version information.
 func Get() (v Version) {
-	date, e := strconv.ParseInt(date, 10, 64)
+	dt, e := strconv.ParseInt(date, 10, 64)
 	if e != nil || len(commit) != 40 {
 		v.Version = "development"
 		v.Commit = "unknown"
@@ -33,7 +40,7 @@ func Get() (v Version) {
 	}
 
 	v.Commit = commit
-	v.Date = time.Unix(date, 0)
+	v.Date = time.Unix(dt, 0)
 	v.Dirty = dirty != ""
 	dirtySuffix := ""
 	if v.Dirty {
@@ -42,13 +49,6 @@ func Get() (v Version) {
 	v.Version = fmt.Sprintf("v0.0.0-%s-%s%s", v.Date.Format("20060102150405"), commit[:12], dirtySuffix)
 	return
 }
-
-// Variables replaced via -ldflags -X.
-var (
-	commit string
-	date   string
-	dirty  string
-)
 
 // GqlVersionType is the GraphQL type for Version struct.
 var GqlVersionType *graphql.Object
